trig tests

+++ scheme/src/read.c
@@ -699,7 +699,7 @@ object sfs_read_real_number(char *input, uint *h) {
         (*h)++;
     }
 
-    double cur_number = 0;
+    long double cur_number = 0;
     do { /* Ca lit la partie entiere chiffre par chiffre */
         if (!isdigit(input[*h])) {
             WARNING_MSG("Invalid number found. \"%c\" is not a valid character in a number",
